Encrypt variables within a data frame

This function corresponds to the in-session version of the encryption process. For the considered loaded dataset, variables to encrypt can be in any number, following users's needs.

crypt_data(
  loaded_dataset,
  vars_to_encrypt,
  vars_to_remove = NULL,
  encryption_key,
  algorithm = "md5",
  correspondence_table = TRUE,
  correspondence_table_label = NULL
)

Arguments

loaded_dataset

The dataset for which user wants to encrypt at least one variable. It must be an expression and not a character vector.

vars_to_encrypt

Character vector of variables to encrypt.

vars_to_remove

Character vector of variables to remove.

encryption_key

Character vector.

algorithm

Algorithm to use. From digest::digest()'s algo argument.

correspondence_table

Logical 1L. TRUE if user wants a correspondence table between initial and encrypted values.

correspondence_table_label

Character 1L. Label of the correspondence table.

See also

Examples

crypt_data(
  loaded_dataset = mtcars[1:5, ], 
  vars_to_encrypt = "mpg", 
  vars_to_remove = "cyl", 
  encryption_key = "1234567", 
  algorithm = "md5", 
  correspondence_table = FALSE
)
#>                                          mpg_crypt disp  hp drat    wt  qsec vs
#> Mazda RX4         CEA9D4F43791ACE22D90EDBF6CE405F5  160 110 3.90 2.620 16.46  0
#> Mazda RX4 Wag     CEA9D4F43791ACE22D90EDBF6CE405F5  160 110 3.90 2.875 17.02  0
#> Datsun 710        65080BE2078D5C5934F018F806E72E81  108  93 3.85 2.320 18.61  1
#> Hornet 4 Drive    18C5941010E3CDCB669F472DF3A1DD5E  258 110 3.08 3.215 19.44  1
#> Hornet Sportabout 1866EA523631A281BD9431915F7C407F  360 175 3.15 3.440 17.02  0
#>                   am gear carb
#> Mazda RX4          1    4    4
#> Mazda RX4 Wag      1    4    4
#> Datsun 710         1    4    1
#> Hornet 4 Drive     0    3    1
#> Hornet Sportabout  0    3    2